backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.110.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
asyncpg==0.29.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
bcrypt==4.1.2

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Data Processing and ML
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
joblib==1.3.2

# Network Analysis and PCAP
scapy==2.5.0
pyshark==0.6
networkx==3.2.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# File Processing
openpyxl==3.1.2
Pillow==10.1.0

# Date and Time
python-dateutil==2.8.2

# Validation and Parsing
email-validator==2.1.0
validators==0.22.0

# Industrial Protocols
pymodbus==3.5.2
# snap7==1.3.0  # Commented out due to compatibility issues

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Additional utilities
click==8.1.7
rich==13.7.0
typer==0.9.0